Rename UploadToPoly to UploadFile
