+++ src/components/dashboard/upload-file/index.js
@@ -1,0 +1,3 @@
+import UploadFile from "./uploadfilecontroller"
+
+export default UploadFile
