--- src/components/dashboard/upload-to-poly/index.js
@@ -1,3 +1,0 @@
-import UploadToPoly from "./uploadtopolycontroller"
-
-export default UploadToPoly
